Reduce cases where root access has to be requested (checking root every time the MainActivity is opened considerably slows down the rendering of the MainActivity):
Add requestRootAccess 'preference' which defaults to true. If root
access is granted once, set it to false. If it fails to run the root
kill commands after the screen turns off, set it to true so that
MainActivity will try to request them again when the user re-opens
the app.